requirements.txt
python-telegram-bot[job-queue]==20.7
psutil==5.9.5
aiohttp>=3.8.0
requests>=2.28.0
python-dateutil>=2.8.2
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
